backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
sqlalchemy>=2.0.22
asyncmy>=0.2.9
alembic>=1.12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
langchain>=0.0.300
langchain-openai>=0.0.2
langchain-community>=0.0.10
chromadb>=0.4.17
matplotlib>=3.8.0
plotly>=5.17.0
pandas>=2.1.1
sentence-transformers>=2.2.2
# MySQL驱动
aiomysql>=0.2.0
pymysql>=1.1.0
# email验证
email-validator>=2.0.0
# 重试机制
tenacity>=8.2.3
aiosqlite>=0.19.0 